+++ Source/General/Misc.vb
@@ -3970,6 +3970,30 @@ Public Enum CommandLinePreview
     <DispName("Windows Terminal")> WindowsTerminal
 End Enum
 
+Public Enum VkResult
+    Success = 0
+    NotReady = 1
+    Timeout = 2
+    EventSet = 3
+    EventReset = 4
+    Incomplete = 5
+    ErrorOutOfHostMemory = -1
+    ErrorOutOfDeviceMemory = -2
+    ErrorInitializationFailed = -3
+    ErrorDeviceLost = -4
+    ErrorMemoryMapFailed = -5
+    ErrorLayerNotPresent = -6
+    ErrorExtensionNotPresent = -7
+    ErrorFeatureNotPresent = -8
+    ErrorIncompatibleDriver = -9
+    ErrorTooManyObjects = -10
+    ErrorFormatNotSupported = -11
+    ErrorFragmentedPool = -12
+    ErrorUnknown = -13
+    ErrorOutOfPoolMemory = -1000069000
+    ErrorInvalidExternalHandle = -1000072003
+End Enum
+
 Public Interface IUpdateUI
     Sub UpdateUI()
 End Interface